fix(critical): mission routes 422 + SLA inactive_agents type error
1. Router registration order: mission/analytics/filter routers BEFORE
   main inspections router. FastAPI matched /inspections/missions against
   /inspections/{inspection_id} (uuid_parsing error on 'missions').
   Fix: register specific prefixes first, catch-all last.

2. Field SLA inactive_agents: CURRENT_DATE - $1 fails because asyncpg
   passes int without implicit cast. Fix: add ::int explicit cast.

+++ packages/backend/app/modules/inspections/services/field_sla_service.py
@@ -439,7 +439,7 @@ class FieldSLAService:
               )
             GROUP BY ap.user_id, u.full_name, u.email, ap.entity_id, e.code
             HAVING MAX(fi.inspection_date) IS NULL
-               OR MAX(fi.inspection_date) < CURRENT_DATE - $1
+               OR MAX(fi.inspection_date) < CURRENT_DATE - $1::int
         """, inactive_days)
 
         if not inactive_rows:
